[Featured] Productos - Add Delete and Update

+++ productos/src/main/java/mingetal/MCM/productos/service/ProductosService.java
@@ -13,7 +13,10 @@ public class ProductosService {
     ProductosRepository productosRepository;
 
     public void save(ProductosEntity productosEntity){
-        productosRepository.save(productosEntity);
+        if(findById(productosEntity.getId())==null){
+            productosRepository.save(productosEntity);
+        }
+
     }
 
     public List<ProductosEntity> findAll(){
@@ -21,7 +24,12 @@ public class ProductosService {
     }
 
     public ProductosEntity findById(int id){
-        return productosRepository.findById(id);
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        return productosEntity;
     }
 
     public List<ProductosEntity> findByTipo(String tipo){
@@ -29,12 +37,68 @@ public class ProductosService {
     }
 
     public ProductosEntity findByNombre(String nombre){
-        return productosRepository.findByNombre(nombre);
+        ProductosEntity productosEntity = productosRepository.findByNombre(nombre);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con nombre " + nombre + " no existe.");
+            return null;
+        }
+        return productosEntity;
     }
 
     public ProductosEntity delete(int id){
         ProductosEntity productosEntity = findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
         productosRepository.delete(productosEntity);
         return productosEntity;
     }
+
+    public ProductosEntity updateValor(int id, int valor){
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        productosEntity.setValor(valor);
+        return productosRepository.save(productosEntity);
+    }
+    public ProductosEntity updateValorFinal(int id, int valor_final){
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        productosEntity.setValor_final(valor_final);
+        return productosRepository.save(productosEntity);
+    }
+    public ProductosEntity updateTipo(int id, String tipo){
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        productosEntity.setTipo(tipo);
+        return productosRepository.save(productosEntity);
+    }
+    public ProductosEntity updateNombre(int id, String nombre){
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        productosEntity.setNombre(nombre);
+        return productosRepository.save(productosEntity);
+    }
+    public ProductosEntity updateCantidad(int id, int cantidad){
+        ProductosEntity productosEntity = productosRepository.findById(id);
+        if(productosEntity==null){
+            //throw new IllegalArgumentException("El producto con ID " + id + " no existe.");
+            return null;
+        }
+        productosEntity.setCantidad(cantidad);
+        return productosRepository.save(productosEntity);
+    }
+
 }
